Rename Complexity_Analysis.pdf to Complexity_Analysis_.pdf
TSP, SAT, and 0-1 Knapsack are three NP-Complete problems with NP- Hard versions Max 0-1 Knapsack, Max SAT, and Min TSP. This report will compare approximation algorithms to solve these problems with the optimal solutions, and will compare the running time of the algorithms.
Max 0-1 Knapsack
For 0-1 Knapsack, there are four algorithms that are compared. Two of them are dynamic programming algorithms that will solve the Knapsack in- stance and return an optimal solution. The Greedy 2-Approximation gives an approximation within a factor of 2 of the optimal, and provides a solution in polynomial time. Lastly, the FPTAS is a fully time polynomial approximation scheme, which is polynomial in the instance size and a factor 1/ε, which is provided to the algorithm along with the instance.
For 0-1 Knapsack, 100 instances of problems are randomly generated and written to a file, which will be read and passed through the algorithms. Important statistics including running time and solution are measured.
SAT
For SAT, the DPLL algorithm will decide the problem and determine whether there exists a solution that solves the SAT instance. A 7-8 Randomized Approx- imation algorithm is implemented for 3-SAT, as well as GSAT, a local search algorithm.
As with 0-1 Knapsack, 100 instances of 3-SAT are randomly generated and written to a file, which will be read and passed through the algorithms. Important statistics including running time and solution are measured.
TSP
For TSP, only one algorithm is implemented to solve it, and it is the Rosenkrantz, Lewis, and Stearns 2-approximation that using Kruskal’s to walk the Minimum Spanning Tree of the STP instance.
The TSP algorithm is tested on TSP instances from TSPLIB, which provides the optimal tour to compare to the approximation.

The rest of the report describes the setup, execution, results, and analysiss of results.